% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{aNewHope}
\alias{aNewHope}
\alias{returnOfTheJedi}
\alias{theEmpireStrikesBack}
\title{aNewHope}
\usage{
aNewHope(...)

theEmpireStrikesBack(...)

returnOfTheJedi(...)
}
\arguments{
\item{...}{\code{swli()} parameters}
}
\description{
Episode specific lorem ipsum
}
\section{Functions}{
\itemize{
\item \code{aNewHope}: Returns Episode 4 content

\item \code{theEmpireStrikesBack}: Returns Episode 5 content

\item \code{returnOfTheJedi}: Returns Episode 6 content
}}
\examples{
aNewHope(paragraphs = 1)
theEmpireStrikesBack(paragraphs = 1)
returnOfTheJedi(paragraphs = 1)
}

