% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{people}
\alias{people}
\alias{places}
\title{people}
\usage{
people(...)

places(...)
}
\arguments{
\item{...}{\code{swli()} parameters}
}
\description{
Star Wars proper nouns
}
\section{Functions}{
\itemize{
\item \code{people}: Returns character names

\item \code{places}: Returns place names
}}
\examples{
people(paragraphs = 1)
places(paragraphs = 1)
}

