% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swli-package.r, R/swli.R
\docType{package}
\name{swli}
\alias{swli}
\alias{swli-package}
\title{swli}
\usage{
swli(paragraphs = 3, type = "e4", src = "installed")
}
\arguments{
\item{paragraphs}{How many paragraphs should be returned}

\item{type}{Provide the episode (3 - 6) or proper noun type
\code{c("e4", "e5", "e6", "places","people")}}

\item{src}{Change default value to pull from github}
}
\value{
swl Star Wars lorem ipsum vector
}
\description{
A lorem ipsum generator using a Star Wars JS library.

Based on the forcem javascript library that generates
Star Wars text from different episodes. The installed
.js file is different to the github version - 
\code{window} was replaced with \code{global}.
}
\details{
When 0 paragraphs are requested, 5 are returned.
}

